# Remove `await` from respond in http_bench.ts